Rename av_decoder module (from decode_av_frames)
